Fix implementation

+++ app/models/Branch.scala
@@ -2,7 +2,7 @@ package models
 
 import play.api.libs.json._
 
-case class Branch(id: Int, name: String, isStable: Boolean, projectId: String)
+case class Branch(id: Long, name: String, isStable: Boolean, projectId: String)
 
 object Branch {
   implicit val format: Format[Branch] = Json.format[Branch]
